+++ src/js/renderers/youtube.js
@@ -1,5 +1,12 @@
 'use strict';
 
+import window from 'global/window';
+import document from 'global/document';
+import mejs from '../core/mejs';
+import {renderer} from '../core/renderer';
+import {createEvent} from '../utils/general';
+import {typeChecks} from '../utils/media';
+
 /**
  * YouTube renderer
  *
@@ -327,7 +334,7 @@ const YouTubeIframeRenderer = {
 									youTubeApi.unMute();
 								}
 								setTimeout(() => {
-									const event = mejs.Utils.createEvent('volumechange', youtube);
+									const event = createEvent('volumechange', youtube);
 									mediaElement.dispatchEvent(event);
 								}, 50);
 								break;
@@ -336,12 +343,12 @@ const YouTubeIframeRenderer = {
 								volume = value;
 								youTubeApi.setVolume(value * 100);
 								setTimeout(() => {
-									const event = mejs.Utils.createEvent('volumechange', youtube);
+									const event = createEvent('volumechange', youtube);
 									mediaElement.dispatchEvent(event);
 								}, 50);
 								break;
 							case 'readyState':
-								const event = mejs.Utils.createEvent('canplay', youtube);
+								const event = createEvent('canplay', youtube);
 								mediaElement.dispatchEvent(event);
 								break;
 
@@ -468,7 +475,7 @@ const YouTubeIframeRenderer = {
 							events = ['mouseover', 'mouseout'],
 							assignEvents = (e) => {
 
-								const newEvent = mejs.Utils.createEvent(e.type, youtube);
+								const newEvent = createEvent(e.type, youtube);
 								mediaElement.dispatchEvent(newEvent);
 							}
 						;
@@ -481,7 +488,7 @@ const YouTubeIframeRenderer = {
 						const initEvents = ['rendererready', 'loadeddata', 'loadedmetadata', 'canplay'];
 
 						for (i = 0, il = initEvents.length; i < il; i++) {
-							const event = mejs.Utils.createEvent(initEvents[i], youtube);
+							const event = createEvent(initEvents[i], youtube);
 							mediaElement.dispatchEvent(event);
 						}
 					},
@@ -537,13 +544,13 @@ const YouTubeIframeRenderer = {
 
 						// send events up
 						for (i = 0, il = events.length; i < il; i++) {
-							const event = mejs.Utils.createEvent(events[i], youtube);
+							const event = createEvent(events[i], youtube);
 							mediaElement.dispatchEvent(event);
 						}
 
 					},
 					onError: (e) => {
-						const event = mejs.Utils.createEvent('error', youtube);
+						const event = createEvent('error', youtube);
 						event.data = e.data;
 						mediaElement.dispatchEvent(event);
 					}
@@ -592,7 +599,7 @@ const YouTubeIframeRenderer = {
 			// create timer
 			youtube.interval = setInterval(() => {
 
-				const event = mejs.Utils.createEvent('timeupdate', youtube);
+				const event = createEvent('timeupdate', youtube);
 				mediaElement.dispatchEvent(event);
 
 			}, 250);
@@ -613,10 +620,10 @@ if (window.postMessage && typeof window.addEventListener) {
 		YouTubeApi.iFrameReady();
 	};
 
-	mejs.Utils.typeChecks.push((url) => {
+	typeChecks.push((url) => {
 		url = url.toLowerCase();
 		return (url.includes('//www.youtube') || url.includes('//youtu.be')) ? 'video/x-youtube' : null;
 	});
 
-	mejs.Renderers.add(YouTubeIframeRenderer);
+	renderer.add(YouTubeIframeRenderer);
 }